Mdttyagi
A swirling vortex, signifying mainnet launch
Pi Network has gained massive attention within the cryptocurrency community. With its unique approach to mining directly from a phone and its potential as a legitimate payment project, Pi Network has accumulated a substantial following.

Yet, the question that lingers in the minds of many is the Pi Network launch date — the long-awaited Open Mainnet when the ecosystem moves from a closed, beta ecosystem into an open network that can be used across exchanges and other blockchains.

Pi was formed in March 2019, and now five years later, the mainnet launch seems imminent. 2024 seems to be the year for the army of ‘Pioneers’ around the world, with just a few key milestones to go.